Moved the parse position into the DirectiveError class.
git-svn-id: https://open-delta.googlecode.com/svn/trunk@1610 dd429c29-c832-59a0-e864-46e68368aa0b

+++ common/src/main/java/au/org/ala/delta/directives/args/ParsingUtils.java
@@ -17,7 +17,7 @@ public class ParsingUtils {
 			return Integer.parseInt(value);
 		}
 		catch (Exception e) {
-			throw DirectiveError.asException(Error.INTEGER_EXPECTED, context.getCurrentOffset());
+			throw DirectiveError.asException(Error.INTEGER_EXPECTED, (int)context.getCurrentOffset());
 		}
 	}
 }
